Run button to own component
